Move View\Template to Template\Factory for a more proper name
